Split multi-game matches into individual games
When there's more than one game in match (i.e., 8 participants) then
there are actually multiple games in progress. Split the match line into
game-sized lists of teams to represent this.

This prevents teams competing in the same match but a different game
from being recorded as facing each other.

The nomenclature for this could be better defined.